attempt to fix CI

+++ src/generator_gemm_avx512_microkernel.c
@@ -2425,7 +2425,7 @@ LIBXSMM_API_INTERN void libxsmm_generator_gemm_avx512_microkernel_fsdbcst( libxs
   if ( LIBXSMM_DATATYPE_BF8 == LIBXSMM_GETENUM_INP( i_xgemm_desc->datatype) ) {
     for ( l_k = 0; l_k < i_k_blocking; l_k++) {
       libxsmm_generator_gemm_avx512_microkernel_bf8_emu_nofsdbcst( io_generated_code, i_gp_reg_mapping, i_micro_kernel_config,
-                                                                   i_xgemm_desc, 1, i_n_blocking, (i_k_blocking == i_xgemm_desc->k) ? l_k : -1 );
+                                                                   i_xgemm_desc, 1, i_n_blocking, (int)((i_k_blocking == i_xgemm_desc->k) ? l_k : -1) );
     }
     return;
   }
